changed some things around related to search
