+++ routes/questions.js
@@ -7,18 +7,33 @@ const { Question } = require('../db/models');
 const { asyncHandler, csrfProtection } = require('./utils');
 const { questionValidators } = require('./validators');
 
+
 const router = express.Router();
+
+// ================================================
+// moved to search.js
+// let searchRepo;
+// let loadingModuleError;
+// try {
+//   searchRepo = require("./search");
+// } catch (e) {
+//   console.error(e);
+//   loadingModuleError = `An error was raised "${e.message}". Check the console for details.`;
+// }
+
+// ================================================
+
 
 //API endpoint for posting/adding a new question
 router.post(
   '/',
   questionValidators,
   asyncHandler(async (req, res) => {
-    const { question, title, categortyId, userId } = req.body;
+    const { question, title, categoryId, userId } = req.body;
     const newQuestion = Question.build({
       question,
       title,
-      categortyId,
+      categoryId,
       userId,
     });
 
@@ -48,12 +63,12 @@ router.post(
     const questionId = parseInt(req.params.id, 10);
     const questionToUpdate = await Question.findByPk(questionId);
 
-    const { question, title, categortyId, userId } = req.body;
+    const { question, title, categoryId, userId } = req.body;
 
     const newQuestion = {
       question,
       title,
-      categortyId,
+      categoryId,
       userId,
     };
 
@@ -88,14 +103,26 @@ router.post(
 
 
 // ================= SEARCH =========================
-// move to search.js
-// async function searchRecipes(term) {
-//   // Use the findAll method of the Question object to search for questions with the
-//   // given term in its title
-//   //
-//   // Docs: https://sequelize.org/v5/manual/querying.html
-//   return await Question.findAll({ where: { title: { [Op.like]: term } } });
-// }
+// moved to search.js
+
+// router.get('/search', async (req, res) => {
+//   let error = loadingModuleError;
+//   let questions;
+//   let answers;
+//   try {
+//     questions = await searchRepo.searchQuestions(`%${req.query.term}%`);
+//     answers = await searchRepo.findAnswer(`%${req.query.term}%`);
+//   } catch (e) {
+//     console.error(e);
+//     error = `An error ocurred that reads "${e.message}". Check the console for more details.`;
+//   }
+//   res.render("search-result.pug", {
+//     listTitle: "Search Results",
+//     error,
+//     questions,
+//     answers,
+//   });
+// });
 
 // ==================================================
