+++ views/search-result.pug
@@ -1,0 +1,36 @@
+extends layout.pug
+
+append head
+    link(rel="stylesheet" href="/stylesheets/index.css")
+    
+include utils
+
+block content
+  h1= listTitle
+  if error
+    +validationErrorSummary(error)
+      p= error
+  else
+    if questions
+      div(class="answer-question-result")
+        each question in questions
+          a.search-link(href="/questions" + question.id)
+              h3.question-title= question.title
+              p= question.question
+        //- each answer in answers
+        //-         a.search-link(href="/answers" + answer.id)
+        //-       h3(class="answer-title")= answer.title
+        //-       p= answer.answer
+    else
+      div(class="need-alert")
+        h3 You have no recipes
+        p Why don't you go and #[a(href="/question/new") make one]!
+//-         //- "p" needs editing
+
+//- //- ==================================================
+//- //- add css below:
+
+//- //- .undecorated,
+//- //- .undecorated:hover {
+//- //-   text-decoration: none;
+//- //- }
